🥖🥅 ↝ Mapped out some potentially problematic component misunderstandings
