+++ components/Content/RoverData/RandomImage.tsx
@@ -1,7 +1,6 @@
 import React, { useState, useEffect } from "react";
 import axios from 'axios';
 import { useSupabaseClient, useSession } from "@supabase/auth-helpers-react";
-import RoverImageCard from "./RoverImageCard";
 import { Form } from "react-hook-form";
 
 const RoverImage = ({ date, rover }) => {
@@ -122,5 +121,29 @@ export function RoverGallerySingle({ inventoryPlanetId }) {
 
     return (
         <RoverImageCard roverImage={{ id: '1', content: '', image: roverImage, planets2: '' }} /> // Form components to be placed in
+    )
+}
+
+interface Props {
+    roverImage: {
+        id: string;
+        content: string;
+        image: string;
+        planets2: string;
+    };
+};
+
+export function RoverImageCard({ roverImage }: Props) {
+    const { id, content, image, planets2 } = roverImage;
+    return (
+        <div className="col-md-4 mb-4">
+            <div style={{ width: '100%', height: "200px", overflow: "hidden" }}>
+                <img
+                    src={image}
+                    alt="Rover Image"
+                    className="w-100 h-auto"
+                />                    
+            </div>
+        </div>
     )
 }
